backend/requirements.txt
anthropic>=0.40.0
openai>=1.30.0
python-telegram-bot>=20.7
websockets>=12.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
aiohttp>=3.9.0

# Mac desktop packaging. Only needed for ctxant_app.py + the onboarding GUI +
# producing ctxant.app. `python main.py` works without them.
rumps>=0.4.0 ; sys_platform == "darwin"
pywebview>=5.0 ; sys_platform == "darwin"
pyinstaller>=6.0 ; sys_platform == "darwin"
